refactor amounts from Double to Long. Add goal/transaction ui refactoring init

+++ composeApp/src/commonMain/kotlin/com/dmytroherez/savingstrack/core/utils/Extensions.kt
@@ -1,6 +1,6 @@
 package com.dmytroherez.savingstrack.core.utils
 
-expect fun formatAsFiat(amount: Double, currencyCode: String, showSign: Boolean = false): String
+expect fun formatAsFiat(amount: Long, currencyCode: String, showSign: Boolean = false): String
 
 object Extensions {
