+++ server/main/kotlin/com/dmytroherez/savingstrack/domain/model/TransactionsAggregatedTotal.kt
@@ -5,5 +5,5 @@ import com.dmytroherez.savingstrack.dto.transactions.SavingCategory
 data class TransactionsAggregatedTotal(
     val category: SavingCategory,
     val currency: String,
-    val sum: Double
+    val sum: Long
 )
